rename NewsletterBrevo to Brevo
